+++ packages/agent-infra/mcp-servers/browser/src/index.ts
@@ -7,12 +7,27 @@
  * Copyright (c) Microsoft Corporation.
  * https://github.com/microsoft/playwright-mcp/blob/main/LICENSE
  */
-import { startSseAndStreamableHttpMcpServer } from 'mcp-http-server';
+import {
+  MiddlewareFunction,
+  startSseAndStreamableHttpMcpServer,
+} from 'mcp-http-server';
 import { program } from 'commander';
+import { BaseLogger, Logger } from '@agent-infra/logger';
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
-import { createServer, getBrowser } from './server.js';
+import { createServer, getBrowser, setConfig, getConfig } from './server.js';
 import { ContextOptions } from './typings.js';
 import { parserFactor, parseViewportSize } from './utils/utils.js';
+import { setRequestContext, getRequestContext } from './request-context.js';
+
+const preActionCallbacks: Array<() => void | Promise<void>> = [];
+const onBeforeStart = (callback: () => void | Promise<void>) => {
+  preActionCallbacks.push(callback);
+};
+
+const middlewares: MiddlewareFunction[] = [];
+const addMiddleware = (middleware: MiddlewareFunction) => {
+  middlewares.push(middleware);
+};
 
 declare global {
   interface Window {
@@ -100,6 +115,15 @@ program
     '--vision',
     'Run server that uses screenshots (Aria snapshots are used by default)',
   )
+  .hook('preAction', async () => {
+    console.log(
+      '[mcp-server-browser] Initializing middlewares and configurations...',
+    );
+
+    for (const callback of preActionCallbacks) {
+      await callback();
+    }
+  })
   .action(async (options) => {
     try {
       console.log('[mcp-server-browser] options', options);
@@ -130,55 +154,21 @@ program
             }),
           },
           contextOptions,
-          logger: {
-            info: (...args: any[]) => {
-              console.log(...args);
-              server.server.notification({
-                method: 'notifications/message',
-                params: {
-                  level: 'warning',
-                  logger: 'mcp-server-browser',
-                  data: JSON.stringify(args),
-                },
-              });
-
-              server.server.sendLoggingMessage({
-                level: 'info',
-                data: JSON.stringify(args),
-              });
-            },
-            error: (...args: any[]) => {
-              console.error(...args);
-              server.server.sendLoggingMessage({
-                level: 'error',
-                data: JSON.stringify(args),
-              });
-            },
-            warn: (...args: any[]) => {
-              console.warn(...args);
-              server.server.sendLoggingMessage({
-                level: 'warning',
-                data: JSON.stringify(args),
-              });
-            },
-            debug: (...args: any[]) => {
-              console.debug(...args);
-              server.server.sendLoggingMessage({
-                level: 'debug',
-                data: JSON.stringify(args),
-              });
-            },
-          },
         });
 
         return server;
       };
       if (options.port || options.host) {
+        const config = getConfig();
         await startSseAndStreamableHttpMcpServer({
           host: options.host,
           port: options.port,
+          middlewares,
+          logger: config.logger,
           // @ts-expect-error: CommonJS and ESM compatibility
           createMcpServer: async (req) => {
+            setRequestContext(req);
+
             const userAgent = req?.headers?.['x-user-agent'] as string;
 
             // header priority: req.headers > process.env.VISION_FACTOR
@@ -213,10 +203,20 @@ program
     }
   });
 
-program.parse();
+process.nextTick(() => {
+  program.parseAsync();
+});
 
 process.stdin.on('close', () => {
   const { browser } = getBrowser();
   console.error('Puppeteer MCP Server closed');
   browser?.close();
 });
+
+export {
+  setConfig,
+  BaseLogger,
+  getRequestContext,
+  onBeforeStart,
+  addMiddleware,
+};