+++ packages/agent-infra/mcp-servers/browser/tests/index.ts
@@ -1,0 +1,27 @@
+#!/usr/bin/env node
+
+import {
+  onBeforeStart,
+  BaseLogger,
+  setConfig,
+  addMiddleware,
+} from '../src/index';
+
+class CustomLogger extends BaseLogger {
+  info(...args) {
+    console.log('custom');
+    console.log(...args);
+  }
+}
+
+onBeforeStart(async () => {
+  addMiddleware((req, res, next) => {
+    console.log('req', req.headers);
+    next();
+  });
+
+  setConfig({
+    logger: new CustomLogger(),
+  });
+  console.log('setConfig');
+});