+++ packages/agent-infra/mcp-servers/browser/CHANGELOG.md
@@ -1,8 +1,54 @@
 # @agent-infra/mcp-server-browser
 
+## 1.2.3
+
+### Patch Changes
+
+- 9a5a933: feat: add custom entry
+- 20aabee: feat: support getRequestContext
+- feat: support logger
+- 0eb2630: feat: add server custom logger
+- Updated dependencies [9a5a933]
+- Updated dependencies [20aabee]
+- Updated dependencies
+- Updated dependencies [0eb2630]
+  - mcp-http-server@1.2.1
+
+## 1.2.3-beta.2
+
+### Patch Changes
+
+- feat: add server custom logger
+- Updated dependencies
+  - mcp-http-server@1.2.1-beta.2
+
+## 1.2.3-beta.1
+
+### Patch Changes
+
+- feat: add custom entry
+- Updated dependencies
+  - mcp-http-server@1.2.1-beta.1
+
+## 1.2.3-beta.0
+
+### Patch Changes
+
+- feat: support getRequestContext
+- Updated dependencies
+  - mcp-http-server@1.2.1-beta.0
+
 ## 1.2.2
 
+### Patch Changes
+
+- chore: support custom logger
+
 ## 1.2.1
+
+### Patch Changes
+
+- chore: support custom logger
 
 ## 1.2.0