+++ packages/agent-infra/mcp-servers/filesystem/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @agent-infra/mcp-server-filesystem
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [9a5a933]
+- Updated dependencies [20aabee]
+- Updated dependencies
+- Updated dependencies [0eb2630]
+  - mcp-http-server@1.2.1
+
+## 1.2.3-beta.2
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.2.1-beta.2
+
+## 1.2.3-beta.1
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.2.1-beta.1
+
+## 1.2.3-beta.0
+
+### Patch Changes
+
+- Updated dependencies
+  - mcp-http-server@1.2.1-beta.0
+
 ## 1.2.2
 
 ## 1.2.1